refactor file structure for triton kernels
