+++ auto_gptq/nn_modules/triton_utils/__init__.py
@@ -1,0 +1,4 @@
+class TritonModuleMixin:
+    @classmethod
+    def warmup(cls, model, transpose=False, seqlen=2048):
+        pass
